refactor: use updated Action type
