+++ src/lib/actions/types.ts
@@ -6,3 +6,12 @@ export interface ActionReturn<Parameter> {
 export interface Action<Parameter = void, Return = ActionReturn<Parameter>> {
 	<Node extends HTMLElement>(node: Node, parameter: Parameter): Return;
 }
+
+export interface ActionReturnAsync<Parameter> {
+	update?: (parameter: Parameter) => Promise<void>;
+	destroy?: () => void;
+}
+
+export interface ActionAsync<Parameter = void, Return = ActionReturnAsync<Parameter>> {
+	<Node extends HTMLElement>(node: Node, parameter: Parameter): Return;
+}
